[tcat] feat: add optional argument to 'present_hash pskd' command
This commit adds an optional 'random_challenge' argument to the 'present_hash pskd' command.
It allows the passing of a custom random challenge as a payload to the 'PRESENT_PSKD_HASH' TLV.

In the current implementation, the random challenge is saved to the variable inside the BBTC code and used by other commands.
In this way, it is not possible to use an arbitrary random challenge value.

This change to the BBTC allows testing of 'present_hash pskd' command with a custom random challenge value.